proceeds into refactoring && QoL fixes 🧚‍♂️
breaks up red.py into two files (actions.py);
protects and provides clear instructions on the input;

problems spotted:
- lack of unicode;
- releases dictionary is not an efficient enough solution: maybe it is
possible to build it through the API?
